fmbtandroid and fmbtlogger Python libraries added
fmbtandroid is a Python library for testing Android devices via
USB. It supports, for instance, verifying, tapping and swiping bitmaps
and UI elements, and shell execution with proper exit status, stdout
and stderr separation.

fmbtlogger is a generic logger that captures method calls, return
values and exceptions of given objects.