test generation & AAL/Python: adapter() changing state case loosened
Changing the state of the model in adapter() blocks of actions is
sometimes useful. This patch enables successful execution of an action
reported from an adapter() if the guard of the action was True before
the adapter() or it became True as the result of executing the
adapter(). Works in AAL/Python, not checked in AAL/C++.