fmbtgti: added screenshot limit and archive method
setScreenshotLimit() defines the maximum number of unused screenshots
that are stored to disk as is. If the limit is exceeded, screenshots
that are already stored to disk will be archived. Currently two archive
methods are supported: removing and resizing.
setScreenshotArchiveMethod() is provided for configuring the method.

This breaks API on an earlier devel version: screenshotLimit is no
more a parameter of enableVisualLog. (Archiving screenshots is needed
also without visual logging.)