fmbtgti: dumpOcr() returns bounding boxes
screenshot().dumpOcr() changed to return list of pairs (recognized word,
bounding box). Bounding box is a quadruple (left, top, bottom, right)
coordinates. dumpOcr() used to return list of triplets (recorgnized
word, x, y) where (x, y) was the upper left coordinate of the word. This
causes an API break on a function that is documented to be used for
debugging purposes.