Fixed vector indexing out of range
This fixes two places where the code tries to index into an empty vector
when operating on a configuration that does not have any tags. I think
the way the code is written it usually works in practice but is
undefined behavior. This commit changes the code so that it does not
even take the address of the first element of the vector unless the
vector is non-empty.

Here is the resulting diff of the generated mycountertest.cc file from
examples/c++-unittest:

--- mycountertest-old.cc	2015-07-25 19:22:29.775713236 -0700
+++ mycountertest-new.cc	2015-07-25 19:23:03.008128216 -0700
@@ -236,7 +236,11 @@
 }
 virtual int getprops(int** props) {
 	tags.clear();
-	*props = &tags[0];
+	if (tags.empty()) {
+		*props = NULL;
+	} else {
+		*props = &tags[0];
+	}
 	return tags.size();
 }
 };