AAL/Python: added --accel-drop-outputs to remote_pyaal
Outputs can be safely dropped from LTS acceleration of lookahead
heuristic test generation if they are not expected to be covered (or
unavoidably on the paths that should be covered) in the test. This is
the case with models where outputs are mostly observations of
exceptional behavior. Dropping outputs may result in significant
performance boost to lookahead test generation.